# service/openapi.yaml
# Generated with protoc-gen-openapi
# https://github.com/google/gnostic/tree/master/cmd/protoc-gen-openapi

openapi: 3.0.3
info:
    title: Gobang API
    description: The greeting service definition.
    version: 0.0.1
paths:
    /helloworld/{name}:
        get:
            tags:
                - Gobang
            operationId: Gobang_SayHello
            parameters:
                - name: name
                  in: path
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/gobang.v1.HelloReply'
    /member/login:
        post:
            tags:
                - Gobang
            operationId: Gobang_Login
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/gobang.v1.LoginReq'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/gobang.v1.LoginReply'
    /member/register:
        post:
            tags:
                - Gobang
            operationId: Gobang_Register
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/gobang.v1.RegisterReq'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/gobang.v1.RegisterReply'
    /member/status:
        get:
            tags:
                - Gobang
            operationId: Gobang_MemberStatus
            parameters:
                - name: sessionid
                  in: query
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/gobang.v1.StatusReply'
    /room/create:
        post:
            tags:
                - Gobang
            operationId: Gobang_RoomCreate
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/gobang.v1.RoomRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/gobang.v1.RoomReply'
components:
    schemas:
        gobang.v1.HelloReply:
            type: object
            properties:
                message:
                    type: string
            description: The response message containing the greetings
        gobang.v1.LoginReply:
            type: object
            properties:
                token:
                    type: string
        gobang.v1.LoginReq:
            type: object
            properties:
                username:
                    type: string
                password:
                    type: string
        gobang.v1.RegisterReply:
            type: object
            properties:
                id:
                    type: integer
                    format: int64
                username:
                    type: string
                lastLoginAt:
                    type: string
                createdAt:
                    type: string
        gobang.v1.RegisterReq:
            type: object
            properties:
                username:
                    type: string
                password:
                    type: string
        gobang.v1.RoomReply:
            type: object
            properties: {}
        gobang.v1.RoomRequest:
            type: object
            properties:
                title:
                    type: string
        gobang.v1.StatusReply:
            type: object
            properties:
                id:
                    type: integer
                    format: int32
                username:
                    type: string
tags:
    - name: Gobang